Support On-premise Jitsi Screen Sharing
The current code that enables Jitsi screen sharing works only with the public Jitsi servers.

This fix uses the preferences field to decide if the js file should be loaded. The default value of the field is the same as the existing string used, so should work fine for public server out of the box.