[FIX] Resolves issue with timing out when Rocket.Chat is in the background … (#1264)
* Resolves issue with timing out when Rocket.Chat is in the background by explicitly setting the awayOnWindowBlur: false instead of relying on the default value

Also, shortened the idleTimeDetectionInterval to 5000 ms. The original value was 1/2 of the idleTimeLimit, which is way too long to change state (e.g. the default of 300 would only check every 150 seconds!

* This makes going from away to online from keyboard/mouse, without Rocket.Chat being focused, work consistently.

This may be too spammy @ 5 seconds, but trigger online/away every time (vs using the cached value) was the only way I could get it consistent.